This is version 7.0.0
This version is a rewrite including new features such as the options file. To change the options in this file, you will need to open it in the Python IDE. Please don't remove or change anything if you don't why it's there. If you do, you are welcome to develop it and make request which I will review.