Fixes and some new features!
Now, there is a brand spanking new Installer! So far, there is theoretical support back to python 3.0. Fixed a typo in JTTools.py. If you find that this code does work on less than 3.0 please open an issue with screenshots of the code working with NO errors and please supply your version number of Python. Thanks for waiting. You now have the update!